move test functions to sub-module, clean up test_exercise
