+++ pythonwhat/test_exercise.py
@@ -2,47 +2,21 @@ from pythonwhat.State import State
 from pythonwhat.utils import check_str, check_dict, check_process
 from pythonwhat.Reporter import Reporter
 from pythonwhat.Test import TestFail
-
-# explicitly import all functions so that they can be used in SCT
-from pythonwhat.test_mc import test_mc
-from pythonwhat.test_or import test_or
-from pythonwhat.test_with import test_with
-from pythonwhat.test_comp import test_list_comp
-from pythonwhat.test_comp import test_dict_comp
-from pythonwhat.test_comp import test_generator_exp
-from pythonwhat.test_import import test_import
-from pythonwhat.test_object import test_object
-from pythonwhat.test_correct import test_correct
-from pythonwhat.test_if_else import test_if_else
-from pythonwhat.test_for_loop import test_for_loop
-from pythonwhat.test_function import test_function
-from pythonwhat.test_function import test_print
-from pythonwhat.test_function import test_function_v2
-from pythonwhat.test_operator import test_operator
-from pythonwhat.test_try_except import test_try_except
-from pythonwhat.test_data_frame import test_data_frame
-from pythonwhat.test_dictionary import test_dictionary
-from pythonwhat.test_while_loop import test_while_loop
-from pythonwhat.test_student_typed import test_student_typed
-from pythonwhat.test_object_accessed import test_object_accessed
-from pythonwhat.test_output_contains import test_output_contains
-from pythonwhat.test_lambda_function import test_lambda_function
-from pythonwhat.test_expression_result import test_expression_result
-from pythonwhat.test_expression_output import test_expression_output
-from pythonwhat.test_function_definition import test_function_definition
-from pythonwhat.test_object_after_expression import test_object_after_expression
+from pythonwhat.sub_test import state_decorator
+from pythonwhat.probe import create_test_probes
 
 # utilities for signatures
+cntxt = {}
+exec("from pythonwhat.test_funcs import *", None, cntxt)
+for t in cntxt: cntxt[t] = state_decorator(cntxt[t])
+
+
+imports = """
 from inspect import Parameter as param
 from pythonwhat.signatures import sig_from_params, sig_from_obj
 from pythonwhat.State import set_converter
-from pythonwhat.probe import create_test_probes
-
-from pythonwhat.sub_test import state_decorator
-
-all_test_names = list(filter(lambda s: s != 'test_exercise' and s.startswith('test'), globals().keys()))
-
-for t in all_test_names: globals()[t] = state_decorator(globals()[t])
+"""
+exec(imports, None, cntxt)
 
 def test_exercise(sct,
                   student_code,
@@ -85,9 +59,9 @@ def test_exercise(sct,
 
     State.set_active_state(state)
     State.TEST_TOP_LEVEL = True
-
-    tree, context = create_test_probes(globals())
-    exec(sct, context)
+    
+    tree, sct_cntxt = create_test_probes(cntxt)
+    exec(sct, sct_cntxt)
 
     # check if no fails yet (can be because of syntax and indentation errors)
     if not rep.failed_test:
@@ -113,3 +87,5 @@ def success_msg(message):
 def allow_errors():
     rep = Reporter.active_reporter
     rep.allow_errors()
+
+cntxt['success_msg'] = success_msg
