+++ pythonwhat/test_funcs/test_correct.py
@@ -2,7 +2,7 @@ from pythonwhat.State import State
 from pythonwhat.Reporter import Reporter
 from pythonwhat.sub_test import sub_test
 
-from pythonwhat.test_or import test_or
+from .test_or import test_or
 
 def test_correct(check, diagnose, state=None):
     rep = Reporter.active_reporter
