+++ pythonwhat/test_funcs/test_for_loop.py
@@ -4,7 +4,7 @@ from pythonwhat.Reporter import Reporter
 from pythonwhat.Test import Test
 from pythonwhat.utils import get_ord
 from pythonwhat.Feedback import Feedback
-from .sub_test import sub_test
+from pythonwhat.sub_test import sub_test
 
 from functools import partial
