+++ pythonwhat/test_funcs/test_object.py
