+++ pythonwhat/test_funcs/test_operator.py
