+++ pythonwhat/test_funcs/test_student_typed.py
